+++ sidecar/proxy/envoy/manager_test.go
@@ -102,7 +102,7 @@ func TestFS(t *testing.T) {
 		},
 	}
 
-	instances := []api.ServiceInstance{
+	instances := []*api.ServiceInstance{
 		{
 			ServiceName: "service1",
 			Endpoint: api.ServiceEndpoint{
@@ -149,6 +149,102 @@ func TestFS(t *testing.T) {
 
 	//err := buildFS(rules)
 	//assert.NoError(t, err)
+}
+
+func TestBuildClusters(t *testing.T) {
+	instances := []*api.ServiceInstance{
+		{
+			ServiceName: "service1",
+			Tags:        []string{"tag1"},
+			ID:          "1",
+			Endpoint:    api.ServiceEndpoint{Type: "http", Value: "127.0.0.1:8080"},
+		},
+		{
+			ServiceName: "service1",
+			Tags:        []string{"tag1", "tag2"},
+			ID:          "2",
+			Endpoint:    api.ServiceEndpoint{Type: "tcp", Value: "127.0.0.5:5050"},
+		},
+		{
+			ServiceName: "service1",
+			ID:          "3",
+			Endpoint:    api.ServiceEndpoint{Type: "tcp", Value: "127.0.0.4:3050"},
+		},
+		{
+			Tags:        []string{"second"},
+			ServiceName: "service2",
+			ID:          "8",
+			Endpoint:    api.ServiceEndpoint{Type: "tcp", Value: "127.0.0.4:3050"},
+		},
+	}
+
+	rules := []api.Rule{
+		{
+			ID:          "abcdef",
+			Destination: "service1",
+			Route: &api.Route{
+				Backends: []api.Backend{
+					{
+						Name:   "service1",
+						Tags:   []string{"tag1"},
+						Weight: 0.25,
+					},
+				},
+			},
+		},
+		{
+			ID:          "abcdef",
+			Destination: "service1",
+			Route: &api.Route{
+				Backends: []api.Backend{
+					{
+						Name:   "service1",
+						Tags:   []string{"tag1", "tag2"},
+						Weight: 0.75,
+						LbType: "random",
+					},
+				},
+			},
+		},
+		{
+			ID:          "abcdef",
+			Destination: "service2",
+			Route: &api.Route{
+				Backends: []api.Backend{
+					{
+						Name:   "service2",
+						Tags:   []string{},
+						Weight: 1.00,
+					},
+				},
+			},
+		},
+		{
+			ID:          "abcdef",
+			Destination: "service2",
+			Actions:     []api.Action{},
+		},
+	}
+
+	clusters := BuildClusters(instances, rules, nil)
+
+	assert.Len(t, clusters, 5)
+
+	clusterName := BuildServiceKey("service1", []string{"tag1"})
+	assert.Equal(t, Cluster{
+		Name:             clusterName,
+		ServiceName:      clusterName,
+		Type:             "sds",
+		LbType:           "round_robin",
+		ConnectTimeoutMs: 1000,
+		OutlierDetection: &OutlierDetection{
+			MaxEjectionPercent: 100,
+		},
+		CircuitBreakers: &CircuitBreakers{},
+	}, clusters[1])
+
+	assert.Equal(t, BuildServiceKey("service1", []string{"tag1", "tag2"}), clusters[2].Name)
+	assert.Equal(t, BuildServiceKey("service2", []string{}), clusters[3].Name)
 }
 
 func TestBuildServiceKey(t *testing.T) {
@@ -272,7 +368,7 @@ func TestBuildParseServiceKey(t *testing.T) {
 }
 
 func TestConvert2(t *testing.T) {
-	instances := []api.ServiceInstance{
+	instances := []*api.ServiceInstance{
 		{
 			ServiceName: "service1",
 			Endpoint: api.ServiceEndpoint{
